requirements.txt
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.32
pydantic==2.8.2
faiss-cpu==1.8.0
google-generativeai==0.7.2
fastapi==0.111.0
uvicorn==0.30.3
python-dotenv==1.0.1
rapidfuzz==3.9.6
protobuf~=4.25.8
requests~=2.32.4
streamlit~=1.48.1